Add workflows for CI/CD
Now there are two workflows to help validate our code automatically on pushes.

The first one is CI, responsible for validating our code on pushes,
independently of branches. The second one, which is CD, runs validations as CI,
and beyond that, it'll push the built code using `clasp`.

For CD, given that it interacts with a project on Google App Scripts, we need to
configure `clasp` for the target project and authenticate with the correct user;
both of those secrets must exist on GitHub Secrets (duh?), respectively, as
`CLASP_ID` and `CLASP_AUTH`. Regarding the values, `CLASP_ID` is just a string,
while `CLASP_AUTH` consists of the contents from `~/.clasprc.json` encoded with
`base64`.